# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/findLimit": {
      "get": {
        "operationId": "AdminController_FindLimitUser",
        "parameters": [
          {
            "name": "take",
            "required": true,
            "in": "query",
            "description": "Number of users to take",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": true,
            "in": "query",
            "description": "Number of users to skip",
            "schema": {
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retrieve a paginated list of users",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/findone/{id}": {
      "get": {
        "operationId": "AdminController_FindUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "9d8c6f2a-ec71-4f24-b5b7-9a21455c2133",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retrieve a user by ID",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/Createone": {
      "post": {
        "operationId": "AdminController_CreateUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "User data object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid user data"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new user",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/Updateone": {
      "put": {
        "operationId": "AdminController_UpdateUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Updated user data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an existing user",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/Deleteone/{id}": {
      "delete": {
        "operationId": "AdminController_DeleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "3b4d4a1e-bc25-4c8d-8f29-11e40a7d8c33",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a user by ID",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/limitProducts": {
      "get": {
        "operationId": "AdminController_FindLimitProducts",
        "parameters": [
          {
            "name": "take",
            "required": true,
            "in": "query",
            "description": "Number of products to take",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": true,
            "in": "query",
            "description": "Number of products to skip",
            "schema": {
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Products retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retrieve a paginated list of products",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/find_one_product/{id}": {
      "get": {
        "operationId": "AdminController_FindProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "c5a3b3f7-987d-47bb-b5dc-81c2d57d1234",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retrieve a product by ID",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/createProduct": {
      "post": {
        "operationId": "AdminController_CreateProduct",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Product data object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid product data"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new product",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/update_Product": {
      "put": {
        "operationId": "AdminController_UpdateProduct",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Updated product data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an existing product",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/delete_Product/{id}": {
      "delete": {
        "operationId": "AdminController_DeleteProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "8d3b7a4f-6a2e-4a1a-9c8a-17c7e8a5f233",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a product by ID",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/find_cart": {
      "get": {
        "operationId": "AdminController_Find_Cart",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "description": "UUID of the entity to find",
            "schema": {
              "example": "b8d5f3e4-8c7d-47a1-b7e5-5a2bca91e7a9",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cart retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          },
          "404": {
            "description": "Cart not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retrieve a cart by ID",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/find_cartItem": {
      "get": {
        "operationId": "AdminController_Find_CartItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "description": "UUID of the entity to find",
            "schema": {
              "example": "2a3f7b1c-4c9d-4b3e-8a2b-11d3f9b8e6f0",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cart item retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartItem"
                }
              }
            }
          },
          "404": {
            "description": "Cart item not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retrieve a cart item by ID",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/getLog": {
      "get": {
        "operationId": "AdminController_getlog",
        "parameters": [
          {
            "name": "take",
            "required": true,
            "in": "query",
            "description": "Number of logs to take",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": true,
            "in": "query",
            "description": "Number of logs to skip",
            "schema": {
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Logs retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retrieve application logs with pagination",
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Portfolio API",
    "description": "API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {}
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "2d9f9c3e-b8c4-4f56-8a2f-23d84b15b9f5",
            "description": "UUID of the user to update"
          },
          "updated_User": {
            "description": "Partial user data to update",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          }
        },
        "required": [
          "id"
        ]
      },
      "Product": {
        "type": "object",
        "properties": {}
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Wireless Mouse",
            "description": "Name of the product"
          },
          "description": {
            "type": "string",
            "example": "High-precision wireless mouse with ergonomic design",
            "description": "Product description"
          },
          "price": {
            "type": "number",
            "example": 49.99,
            "description": "Product price"
          },
          "stock": {
            "type": "number",
            "example": 100,
            "description": "Available stock quantity"
          },
          "more_description": {
            "type": "string",
            "example": "Extra info about materials or specifications",
            "description": "Additional product details"
          },
          "categoryId": {
            "type": "string",
            "example": "electronics",
            "description": "Category ID or name (optional)"
          },
          "images": {
            "example": [
              "image1.jpg",
              "image2.jpg"
            ],
            "description": "Array of product image URLs",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "description",
          "price",
          "stock"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "9e8a1b64-3f24-4c54-ae13-cc1fd6d8a6f0",
            "description": "UUID of the product to update"
          },
          "updatedProduct": {
            "description": "Partial product data to update",
            "allOf": [
              {
                "$ref": "#/components/schemas/Product"
              }
            ]
          }
        },
        "required": [
          "id"
        ]
      },
      "Cart": {
        "type": "object",
        "properties": {}
      },
      "CartItem": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}